refactor(config): 统一配置规则并拆分管理页面
集中配置类型、默认值、校验、迁移和局部更新逻辑，修复零值语义、主题参数保存及损坏配置恢复。

拆分配置页功能区块，补充中文开发规范、41 个回归测试与 GitHub Actions 质量检查。

验证：pnpm check（853 个测试）、隔离生产构建及浏览器保存回读、草稿保留和错误重试。
