+++ .node-version
@@ -1,0 +1,1 @@
+24.16.0
