+++ components/admin/config/storage-settings.tsx
@@ -1,0 +1,31 @@
+"use client"
+
+import { TabsContent } from "@/components/ui/tabs"
+import { Field, FieldGroup, FieldLabel } from "@/components/ui/field"
+import { Input } from "@/components/ui/input"
+import { SectionCard } from "@/components/admin/section-card"
+
+import type { ConfigForm } from "./use-config-form"
+
+export function StorageSettings({
+  cfg,
+  updateField,
+}: Pick<ConfigForm, "cfg" | "updateField">) {
+  return (
+    <TabsContent value="storage">
+      <SectionCard title="存储" description="数据库文件路径。">
+        <FieldGroup>
+          <Field>
+            <FieldLabel htmlFor="dbPath">数据库路径</FieldLabel>
+            <Input
+              id="dbPath"
+              value={cfg.dbPath}
+              placeholder="./data/agent.db"
+              onChange={(e) => updateField("dbPath", e.target.value)}
+            />
+          </Field>
+        </FieldGroup>
+      </SectionCard>
+    </TabsContent>
+  )
+}
