+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.0.1",
   "type": "module",
   "private": true,
+  "packageManager": "pnpm@11.17.0",
   "scripts": {
     "dev": "next dev -H 0.0.0.0",
     "build": "next build",
